+++ main/java/com/multi/shoes4jo/vo/KeywordTrendVO.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Component;
 public class KeywordTrendVO {
 	private String period_sdata;
 	private String keyword;
-	private long ratio_cnt;
+	private int ratio_cnt;
 	
 	public KeywordTrendVO() {
 	}
@@ -29,10 +29,10 @@ public class KeywordTrendVO {
 	public void setKeyword(String keyword) {
 		this.keyword = keyword;
 	}
-	public long getRatio_cnt() {
+	public int getRatio_cnt() {
 		return ratio_cnt;
 	}
-	public void setRatio_cnt(long ratio_cnt) {
+	public void setRatio_cnt(int ratio_cnt) {
 		this.ratio_cnt = ratio_cnt;
 	}
